Rename global_name to constants
