+++ src/sdx/pce/utils/random_topology_generator.py
@@ -14,7 +14,7 @@ import pylab as plt
 from networkx.algorithms import approximation as approx
 from networkx.generators.random_graphs import erdos_renyi_graph
 
-import sdx.pce.utils.global_name as global_name
+import sdx.pce.utils.constants as global_name
 from sdx.pce.utils.functions import GraphFunction
