+++ tests/test_te_solver.py
@@ -3,7 +3,7 @@ import unittest
 
 import networkx as nx
 
-import sdx.pce.utils.global_name as global_name
+import sdx.pce.utils.constants as global_name
 from sdx.pce.load_balancing.te_solver import TE_Solver
 from sdx.pce.utils.random_connection_generator import RandomConnectionGenerator
 from sdx.pce.utils.random_topology_generator import RandomTopologyGenerator, dot_file
